Refactor from express filters
