+++ server.js
@@ -8,6 +8,7 @@ const app            = express();
 const port           = 5001;
 var MongoStore = require('connect-mongo');
 var jwt = require('jsonwebtoken');
+let filters = require('./src/filters');
 
 MongoClient.connect(db.url, function(err, database) {
     if (err) return console.log(err);
@@ -18,22 +19,4 @@ MongoClient.connect(db.url, function(err, database) {
 });
 
 app.use(bodyParser.urlencoded({ extended: true }));
-
-app.use(function(req, res, next) {
-    var unsecurePaths = ['/login', '/signup'];
-    if (unsecurePaths.includes(req.originalUrl)) {
-        next()
-    } else {
-        if (req. headers && req.headers.authorization && req.headers.authorization.split(' ')[0] === 'JWT') {
-            jwt.verify(req.headers.authorization.split(' ')[1], 'mypass', function(err, decode) {
-                if (err) {
-                    res.status(403).send("Invalid user");
-                }
-                req.user = decode;
-                next();
-            });
-        } else {
-            res.status(403).send("Invalid user");
-        }
-    }
-});
+app.use(filters.authentication);
